Move PrimitiveType to file types
