+++ src/builder.ts
@@ -1,10 +1,11 @@
 import { Database, Table } from './database'
 import { BooleanColumn, Column } from './columns'
+import { PrimitiveType } from './models/types'
 import { Condition } from './models/Condition'
 import { Binder, BinderStore } from './binder'
 import { ASTERISK, Distinct, All } from './singletoneConstants'
 import { SelectStep, FromStep, RootStep } from './steps/stepInterfaces'
-import { Step, PrimitiveType, SelectItem } from './steps/Step'
+import { Step, SelectItem } from './steps/Step'
 import { LogicalOperator } from './operators'
 import { Parenthesis } from './steps/BaseStep'
 import { OrderByItemInfo } from './orderBy'
